requirements.txt
# LangGraph Job Agent Bot - Dependencies

# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0

# Telegram Bot
python-telegram-bot>=21.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.0.0

# Job API & HTTP
aiohttp>=3.9.0
requests>=2.31.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
